Bug Fixing 2
Scrutinize should say correct HP/text now.
MP regen cut down to 1/16 max MP.
Disco Hooligan text should have apostrophes.
Frenzied Invasion should be fixed.
Unit flash bug should be fixed

+++ ArchRPG/Assets/Scripts/Battle/Ability.cs
@@ -1971,7 +1971,7 @@ namespace PlayerAbilities
         public override string OutputText(unit user, unit target)
         {
             string output = "";
-            float percent = target.getHP() / target.maxHP;
+            float percent = (float)target.getHP() / (float)target.maxHP;
 
             if (percent > .7f)
             {
